test: refactor karma tests to use test harness
